Implement endpoint to fetch all polygon geofences coordinates.

+++ Fleet ManagementWebApplication/Models/PolygonGeofence.cs
@@ -4,7 +4,7 @@
     {
         public int ID { get; set; }
         public int GeofenceID { get; set; }
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
     }
 }
